Update in Groupme!
Hey Everybody! I just wanted to post an update, I'm about to push a commit through to main, here's a list of everything I've done over the past day that will help you guys
1) Spawnpoint system - just attach a node named spawnpoint to your scene and it will become the player's spawnpoint
2)
6:08 PM
Avatar
Ethan Shepher
AH! I accidentally sent that message to early, here's the rest of it
2) I added a save system that works globally, @Barbara Guerrero all you have to do once you recieve the player's input for their name, is call SaveManager.assignNewPlayer("playername")
check out the savemanager script to see other functions you can access globally

3)save files can have multiple players and will load in based on current player

4)level unlock system : based on the data in the save file it will grey out a button in the menu and you can no longer access it
(has to be three parts sorry)

5) the menu's are all connected so when you hit play it will run the main menu which loads the map which loads the level you suggest

(Why this one is important to know) when adding a level, the naming convention should be Level_Number so like Level_1, because that is how it interacts with the map

6) the map works! (and scales with viewport so if you go fullscreen

7) level 3 is done! Make sure to pull these changes so when you add your level you can access things like the save file or map

In progress:
A level finish checkpoint, that saves your progress and boots you to the map!